separates route and controller logic
